requirements.txt
Django==1.10.*
dj-database-url==0.4.1
gunicorn==19.6.0
whitenoise==3.3.0
setuptools-scm-1.15.5
django-haystack-2.6.0
whoosh-2.7.4
fabric==*
